requirements.txt
fastapi
uvicorn
dnspython
pydantic
email-validator